Move VDiff related workflow server APIs to vdiff.go and add unit tests
